# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YWC API",
        "description": "API documentation for YWC platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://console-ywc.nativeappdev.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/itinerary/routes/{route}/crew": {
            "get": {
                "tags": [
                    "Itinerary Crew"
                ],
                "summary": "Get crew members for a route",
                "operationId": "74c3a6b8c3172ca77c75dd9aba9b6e91",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itinerary Crew"
                ],
                "summary": "Invite crew member to route",
                "operationId": "034577036729440b8c34efa06bee16eb",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email (required if user_id not provided)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "crew@example.com"
                                    },
                                    "user_id": {
                                        "description": "User ID (required if email not provided)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "owner",
                                            "editor",
                                            "viewer"
                                        ],
                                        "example": "editor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crew invitation sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew invitation sent."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itinerary/routes/{route}/crew/{crew}": {
            "put": {
                "tags": [
                    "Itinerary Crew"
                ],
                "summary": "Update crew member",
                "operationId": "1908189fa447957bcf809709c176f984",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crew",
                        "in": "path",
                        "description": "Crew member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "owner",
                                            "editor",
                                            "viewer"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "accepted",
                                            "declined",
                                            "revoked"
                                        ]
                                    },
                                    "notify_on_updates": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crew member updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew member updated."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Crew member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Itinerary Crew"
                ],
                "summary": "Remove crew member",
                "operationId": "3f6e4838c0f4c3e99422a9e0f13683cb",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crew",
                        "in": "path",
                        "description": "Crew member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crew member removed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew member removed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Crew member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itinerary/routes/{route}/crew/{crew}/respond": {
            "post": {
                "tags": [
                    "Itinerary Crew"
                ],
                "summary": "Respond to crew invitation",
                "operationId": "c1e22bd00cd7ab38033afafda3642bd0",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crew",
                        "in": "path",
                        "description": "Crew member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "accept",
                                            "decline"
                                        ],
                                        "example": "accept"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation response processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have joined the crew."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Crew invitation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}